+++ schema/src/main/scala/net/reactivecore/cjs/Location.scala
@@ -4,7 +4,7 @@ import io.circe.syntax._
 import io.circe.{Codec, Decoder, Encoder, Json, JsonObject}
 import net.reactivecore.cjs.resolver.RefUri
 import net.reactivecore.cjs.util.Codecs
-import net.reactivecore.cjs.validator.ValidationProvider
+import net.reactivecore.cjs.validator.provider.ValidationProvider
 
 /** Describes location of a Schema. */
 case class Location(
