+++ schema/src/main/scala/net/reactivecore/cjs/restriction/ConstRestriction.scala
@@ -1,7 +1,8 @@
 package net.reactivecore.cjs.restriction
 
 import io.circe._
-import net.reactivecore.cjs.validator.{ConstValidator, ValidationProvider, Validator}
+import net.reactivecore.cjs.validator.provider.ValidationProvider
+import net.reactivecore.cjs.validator.{ConstValidator, Validator}
 
 /**
   * Restricts to a constant value.
