+++ schema/src/main/scala/net/reactivecore/cjs/restriction/LogicRestrictions.scala
@@ -5,6 +5,7 @@ import io.circe.generic.semiauto
 import net.reactivecore.cjs.resolver.{JsonPointer, RefUri}
 import net.reactivecore.cjs.util.Codecs
 import net.reactivecore.cjs.validator._
+import net.reactivecore.cjs.validator.provider.ValidationProvider
 import net.reactivecore.cjs.{Schema, SchemaOrigin}
 
 /** Further restrictions of a type. */
