+++ schema/src/main/scala/net/reactivecore/cjs/restriction/ObjectRestriction.scala
@@ -3,9 +3,10 @@ package net.reactivecore.cjs.restriction
 import io.circe.Codec
 import io.circe.generic.semiauto
 import net.reactivecore.cjs.util.{Codecs, VectorMap}
-import net.reactivecore.cjs.validator.{ValidationProvider, Validator}
+import net.reactivecore.cjs.validator.Validator
 import net.reactivecore.cjs.validator.obj._
 import net.reactivecore.cjs.Schema
+import net.reactivecore.cjs.validator.provider.ValidationProvider
 
 import java.util.regex.Pattern
