+++ schema/src/main/scala/net/reactivecore/cjs/restriction/SingleRestriction.scala
@@ -2,7 +2,12 @@ package net.reactivecore.cjs.restriction
 
 import io.circe.{Codec, Decoder, Encoder, Json}
 import net.reactivecore.cjs.SchemaOrigin
-import net.reactivecore.cjs.validator.{SimpleContextFreeValidator, ValidationProvider, Validator}
+import net.reactivecore.cjs.validator.provider.{
+  ContextValidationProvider,
+  SequenceValidationProvider,
+  ValidationProvider
+}
+import net.reactivecore.cjs.validator.{SimpleContextFreeValidator, Validator}
 import shapeless._
 import shapeless.labelled.FieldType
 
@@ -10,7 +15,6 @@ import shapeless.labelled.FieldType
 case class SingleRestriction[T, VF](value: T) extends AnyVal
 
 object SingleRestriction {
-
   case class MinimumValidator(value: BigDecimal) extends SimpleContextFreeValidator("minimum") {
     override def isValid(json: Json): Boolean = {
       json.as[BigDecimal] match {
@@ -21,34 +25,11 @@ object SingleRestriction {
     }
   }
 
-  /** Context specific validation provider (context may be the outlying class) */
-  trait ContextValidationProvider[T, -C] {
-    def apply(context: C): ValidationProvider[T]
-  }
-
   /** Provides a validation provider for Validators with trivial constructor */
   implicit def singleValidationProvider[T, V <: Validator](
       implicit generic: Generic.Aux[V, T :: HNil]
   ): ValidationProvider[SingleRestriction[T, V]] = { (origin, value) =>
     generic.from(value.value :: HNil)
-  }
-
-  implicit def fromContextProvider[T](implicit p: ValidationProvider[T]): ContextValidationProvider[T, Any] =
-    (_: Any) => p
-
-  /** Support for optionals */
-  implicit def optionalSupport[T, C](
-      implicit underlying: ContextValidationProvider[T, C]
-  ): ContextValidationProvider[Option[T], C] = {
-    (context: C) =>
-      { (origin: SchemaOrigin, restriction: Option[T]) =>
-        {
-          restriction match {
-            case Some(value) => underlying(context)(origin, value)
-            case None        => Validator.success
-          }
-        }
-      }
   }
 
   /** JSON Codec Support. */
@@ -67,86 +48,19 @@ object SingleRestriction {
   // Das geht auch
   implicitly[Codec[SingleRestriction[BigDecimal, MinimumValidator]]]
 
-  // implicit val trivial = makeTrivialProvider(MinimumValidator.apply)
-
   case class Example(
       minimum: Option[SingleRestriction[BigDecimal, MinimumValidator]]
   )
 
   val codec = io.circe.generic.semiauto.deriveCodec[Example]
 
+  implicitly[ValidationProvider[SingleRestriction[BigDecimal, MinimumValidator]]]
+  implicitly[ContextValidationProvider[SingleRestriction[BigDecimal, MinimumValidator], Example]]
+
   // Jetzt fehlt mir nur noch ein ValidationProvider
 
-  trait SequenceValidationProviderBuilder[C] {
-    def apply: ValidationProvider[C]
-  }
+  val lg = LabelledGeneric[Example]
+  val helper = implicitly[SequenceValidationProvider.Helper[lg.Repr, Example]]
 
-  object SequenceValidationProviderBuilder {
-
-    /** Recursive Helper for Building. */
-    trait Helper[T, C] {
-      def apply(context: C): ValidationProvider[T]
-    }
-
-    implicit def hnilHelper[C]: Helper[HNil, C] = (_: C) => ValidationProvider.empty
-
-    implicit def hlistHelper[K <: Symbol, H, T <: HList, V <: Validator, C](
-        implicit p: ContextValidationProvider[H, C],
-        tailHelper: Helper[T, C],
-        w: Witness.Aux[K]
-    ): Helper[FieldType[K, H] :: T, C] =
-      new Helper[FieldType[K, H] :: T, C] {
-        val fieldName = w.value.name
-
-        override def apply(context: C): ValidationProvider[FieldType[K, H] :: T] = {
-          ValidationProvider.withOrigin { (origin, value) =>
-            Validator.sequence(
-              p.apply(context)(origin.enterObject(fieldName), value.head),
-              tailHelper.apply(context)(origin, value.tail)
-            )
-          }
-        }
-      }
-
-    implicit def generate[C, G](
-        implicit labelledGeneric: LabelledGeneric.Aux[C, G],
-        helper: Helper[G, C]
-    ): SequenceValidationProviderBuilder[C] = {
-      new SequenceValidationProviderBuilder[C] {
-        override def apply: ValidationProvider[C] = {
-          new ValidationProvider[C] {
-            override def apply(origin: SchemaOrigin, restriction: C): Validator = {
-              helper.apply(restriction).apply(origin, labelledGeneric.to(restriction))
-            }
-          }
-        }
-      }
-    }
-  }
-
-  def makeSequenceValidationProvider[C](
-      implicit builder: SequenceValidationProviderBuilder[C]
-  ): ValidationProvider[C] = {
-    builder.apply
-  }
-
-  implicitly[LabelledGeneric[Example]]
-
-  SequenceValidationProviderBuilder.hnilHelper[Example]
-
-  /*
-  SequenceValidationProviderBuilder.hlistHelper[
-    Option[SingleRestriction[BigDecimal, MinimumValidator]],
-    HNil,
-    Example
-  ]
-   */
-
-  /*
-  implicitly[
-    SequenceValidationProviderBuilder.Helper[Option[SingleRestriction[BigDecimal, MinimumValidator]] :: HNil, Example]
-  ]
-   */
-
-  makeSequenceValidationProvider[Example]
+  SequenceValidationProvider.apply[Example]
 }
